requirements.txt
# requirements.txt
streamlit>=1.30.0
python-docx>=0.8.11
deep-translator>=1.11.4
requests>=2.31.0